pyproject.toml
[project]
name = "production_planner"
dynamic = ["version"]
authors = [
    {name = "mitaa", email = "mitaa.ceb@gmail.com"}
    ]
description = "A production planner for Satisfactory"
readme = "README.md"
license= {file = "LICENSE"}
requires-python = ">=3.10"
classifiers = [
        "Programming Language :: Python :: 3",
        "Development Status :: 3 - Alpha",
        "License :: OSI Approved :: MPL 2 License",
        "Operating System :: OS Independent",
    ]
dependencies = [
    "textual[syntax] >= 0.79",
    "rich",
    "platformdirs",
    "pyaml",
    "json-store >= 4.0",
    "docopt >= 0.6.0"
    ]

[project.optional-dependencies]
dev = [
    "pytest >= 8.3.1",
    "pytest-xdist >= 3.6",
    "pytest-asyncio",
    "pytest-cov >= 5.0.0",
    "pytest-textual-snapshot >= 1.0"
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
asyncio_default_fixture_loop_scope = "function"

[build-system]
requires = ["setuptools>=64", "setuptools-scm>=8"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
include-package-data = true

[tool.setuptools_scm]


[tool.setuptools.packages.find]
where = ["src"]

[project.scripts]
production_planner = "production_planner:main"
